#95: WIP testing images add file support to load ans save  jitter dims for tests
