+++ src/vt-tv/render/render.cc
@@ -51,6 +51,20 @@
 
 namespace vt { namespace tv {
 
+void Render::initJitterDims() {
+  std::srand(std::time(nullptr));
+  auto const& allObjects = info_.getAllObjectIDs();
+  for (auto const& objectID : allObjects) {
+    std::array<double, 3> jitterDims;
+    for (uint64_t d = 0; d < 3; d++) {
+      if (auto f = rank_dims_.find(d); f != rank_dims_.end()) {
+        jitterDims[d] = ((double)std::rand()/RAND_MAX - 0.5) *  object_jitter_;
+      } else jitterDims[d] = 0;
+    }
+    jitter_dims_.insert(std::make_pair(objectID, jitterDims));
+  }
+};
+
 Render::Render(Info in_info)
 : info_(in_info) // std:move ?
 , n_ranks_(in_info.getNumRanks())
@@ -94,41 +108,7 @@ Render::Render(Info in_info)
   std::srand(std::time(nullptr));
   auto const& allObjects = info_.getAllObjectIDs();
 
-  jitter_dims_ = {};
-
-#ifndef VT_TV_HAS_TESTS
-  // Load object jitter from file
-  if (std::getenv("VT_TV_OBJECT_JITTER_DIMS_FILE") != "") {
-    // to specify values from file
-    std::ifstream infile(std::getenv("VT_TV_OBJECT_JITTER_DIMS_FILE"));
-    int objectId, x, y, z;
-    while (infile >> objectId >> x >> y >> z)
-    {
-      std::array<double, 3> jitterDims = {static_cast<double>(x), static_cast<double>(y), static_cast<double>(z)};
-      jitter_dims_.insert(std::make_pair(objectId, jitterDims));
-    }
-  }
-#endif
-
-  if (jitter_dims_.empty()) {
-    for (auto const& objectID : allObjects) {
-      auto jitterDims = getObjectJitterDims(objectID, object_jitter_, rank_dims_);
-      jitter_dims_.insert(std::make_pair(objectID, jitterDims));
-    }
-
-#ifndef VT_TV_HAS_TESTS
-  // Save object jitter to file
-  if (std::getenv("VT_TV_OBJECT_JITTER_DIMS_FILE") != "") {
-    // to specify values from file
-    std::ofstream outfile(std::getenv("VT_TV_OBJECT_JITTER_DIMS_FILE"));
-    int objectId, x, y, z;
-    for (auto const& objectID : allObjects) {
-      auto jitterDims = jitter_dims_.at(objectID);
-      outfile << objectId << x << y << z;
-    }
-  }
-#endif
-  }
+  initJitterDims();
 
   object_qoi_range_ = this->computeObjectQoiRange_();
   rank_qoi_range_ = this->computeRankQoiRange_();
@@ -186,32 +166,12 @@ Render::Render(
   }
 
   // Initialize jitter
-  std::srand(std::time(nullptr));
-  auto const& allObjects = info_.getAllObjectIDs();
-  for (auto const& objectID : allObjects) {
-    auto jitterDims = getObjectJitterDims(objectID, object_jitter_, rank_dims_);
-    jitter_dims_.insert(std::make_pair(objectID, jitterDims));
-  }
-#ifdef VT_TV_TESTS
-#endif
+  initJitterDims();
 
   object_qoi_range_ = this->computeObjectQoiRange_();
   rank_qoi_range_ = this->computeRankQoiRange_();
   object_volume_max_ = this->computeMaxObjectVolume_();
   object_load_max_ = this->info_.getMaxLoad();
-};
-
-/**
- * \brief Function to initialize object dimensions using some randomness
- */
-auto getObjectJitterDims(const ElementIDType &objectID, double object_jitter, std::set<uint64_t> rank_dims_) {
-  std::array<double, 3> jitterDims;
-  for (uint64_t d = 0; d < 3; d++) {
-    if (auto f = rank_dims_.find(d); f != rank_dims_.end()) {
-      jitterDims[d] = ((double)std::rand()/RAND_MAX - 0.5) *  object_jitter;
-    } else jitterDims[d] = 0;
-  }
-  return jitterDims;
 };
 
 double Render::computeMaxObjectVolume_() {
