+++ tests/config/ccm-example-no-png.yaml
@@ -1,0 +1,18 @@
+input:
+  directory: tests/data/ccm_example
+  n_ranks: 2
+
+viz:
+  x_ranks: 2
+  y_ranks: 1
+  z_ranks: 1
+  object_jitter: 0.5
+  rank_qoi: load
+  object_qoi: shared_block_id
+  save_meshes: true
+  save_pngs: false
+  force_continuous_object_qoi: true
+
+output:
+  directory: output/tests/no_png
+  file_stem: ccm_example
