+++ tests/config/conf-no-png.yaml
@@ -1,0 +1,18 @@
+input:
+  directory: tests/data/lb_test_data
+  n_ranks: 4
+
+viz:
+  x_ranks: 2
+  y_ranks: 2
+  z_ranks: 1
+  object_jitter: 0.5
+  rank_qoi: load
+  object_qoi: load
+  save_meshes: true
+  save_pngs: false
+  force_continuous_object_qoi: true
+
+output:
+  directory: output/tests/no_png
+  file_stem: default
