+++ tests/unit/render/test_parse_render.cc
@@ -68,186 +68,46 @@ class ParseRenderTest :public ::testing::TestWithParam<std::string> {
 
   void SetUp() override {
     // Make the output directory for these tests
-    std::filesystem::create_directory(fmt::format("{}/output", SRC_DIR));
-    std::filesystem::create_directory(fmt::format("{}/output/tests", SRC_DIR));
-  }
-
-protected:
-
-  void printVtkPolyData(vtkPolyData *poly) {
-    std::string points_str = "";
-    for(auto k=0; k < poly->GetNumberOfPoints(); ++k) {
-      double* coords = poly->GetPoint(k);
-      points_str = fmt::format("{}    [{}] = [{}, {}, {}]\n", points_str, k, coords[0], coords[1], coords[2]);
-    }
-
-    std::string point_array_str = "";
-    for(auto k=0; k < poly->GetPointData()->GetNumberOfArrays(); ++k) {
-      auto array_name = poly->GetPointData()->GetArrayName(k);
-      auto array = poly->GetPointData()->GetArray(array_name);
-
-      std::string components_str = "";
-      for (auto tupleIdx = 0; tupleIdx < array->GetNumberOfTuples(); ++tupleIdx) {
-        for (auto compIdx = 0; compIdx < array->GetNumberOfComponents(); ++compIdx) {
-          components_str = fmt::format("{}    [{}] = [TupleId={}, ComponentId={}, {}]\n", components_str, k, tupleIdx, compIdx);
-        }
-      }
-
-      point_array_str = fmt::format("{}    [{}] = { Name=\"{}\", Type={}, Size={}, Components={} }\n",
-        point_array_str,
-        k,
-        array_name,
-        array->GetArrayType(),
-        array->GetSize(),
-        components_str
-      );
-    }
-
-    fmt::print(
-      R"STR(
-> PointData.NumberOfArrays => {}
-> NumberOfPoints => {}
-> Lines.Data.Name => {}
-> Points => [
-  {}
-  ]
-  PointData.Arrays() => [
-  {}
-  ]
-)STR",
-      
-      poly->GetPointData()->GetNumberOfArrays(),
-      poly->GetNumberOfPoints(),
-      Util::formatNullable(poly->GetLines()->GetData()->GetName()),
-      points_str,
-      point_array_str
-    );
-  }
-
-  void assertPolyEquals(vtkPolyData *actual, vtkPolyData *expected) {
-    fmt::print("Actual vtkPolyData:\n");
-    printVtkPolyData(actual);
-    fmt::print("Expected vtkPolyData:\n");
-    printVtkPolyData(expected);
-
-    // Assertions required to test vt-tv meshaes
-    // Number of point data should be ranks
-    ASSERT_EQ(actual->GetPointData()->GetNumberOfArrays(), expected->GetPointData()->GetNumberOfArrays());
-
-    // Validate points
-    ASSERT_EQ(actual->GetNumberOfPoints(), expected->GetNumberOfPoints());
-
-    for(auto k=0; k < actual->GetNumberOfPoints(); ++k) {
-      double actualCoords[3];
-      double expectedCoords[3];
-      actual->GetPoint(k, actualCoords);
-      expected->GetPoint(k, expectedCoords);
-
-      ASSERT_EQ(actualCoords[0], expectedCoords[0]) << "Invalid point X coordinate at index " << k << "diff=" << (expectedCoords[0] - actualCoords[0]);
-      ASSERT_EQ(actualCoords[1], expectedCoords[1]) << "Invalid point Y coordinate at index " << k << "diff=" << (expectedCoords[1] - actualCoords[1]);
-      ASSERT_EQ(actualCoords[2], expectedCoords[2]) << "Invalid point Z coordinate at index " << k << "diff=" << (expectedCoords[2] - actualCoords[2]);
-    }
-
-    // Validate lines
-    ASSERT_EQ(actual->GetLines()->GetData()->GetName(), expected->GetLines()->GetData()->GetName());
-    // TODO: browse lines
+    // std::filesystem::create_directory(fmt::format("{}/output", SRC_DIR));
+    // std::filesystem::create_directory(fmt::format("{}/output/tests", SRC_DIR));
+    // std::filesystem::create_directory(fmt::format("{}/output/tests/no_png", SRC_DIR));
   }
 };
 
 /**
  * Test ParseRender:parseAndRender correcty run the different configuration files
  */
-TEST_P(ParseRenderTest, test_parse_config_and_render_output) {
+TEST_P(ParseRenderTest, test_parse_config_and_render_no_png) {
   std::string const & config_file = GetParam();
   auto parse_render = ParseRender(fmt::format("{}/tests/config/{}", SRC_DIR, config_file));
-  ASSERT_NO_THROW(parse_render.parseAndRender());
-
   YAML::Node config = YAML::LoadFile(fmt::format("{}/tests/config/{}", SRC_DIR, config_file));
   Info info = Generator::loadInfoFromConfig(config);
 
   auto output_dir = Util::resolveDir(SRC_DIR, config["output"]["directory"].as<std::string>(), true);
+  std::filesystem::create_directories(output_dir);
+
   auto output_file_stem = config["output"]["file_stem"].as<std::string>();
-  // auto n_ranks = config["input"]["n_ranks"].as<int64_t>();
+
+  // generate file containing object jitter dimensions for output trace and input usage
+  // this is needed for image and mesh tests
+  ASSERT_NO_THROW(parse_render.parseAndRender());
 
   for (uint64_t i = 0; i<info.getNumPhases(); i++) {
     auto rank_mesh_file = fmt::format("{}{}_rank_mesh_{}.vtp", output_dir, output_file_stem, i);
     auto object_mesh_file = fmt::format("{}{}_object_mesh_{}.vtp", output_dir, output_file_stem, i);
     auto png_file = fmt::format("{}{}{}.png", output_dir, output_file_stem, i);
-    auto expected_rank_mesh_file = fmt::format("{}/tests/expected/{}/{}_rank_mesh_{}.vtp",
-                                                SRC_DIR, output_file_stem, output_file_stem, i);
-    auto expected_object_mesh_file = fmt::format("{}/tests/expected/{}/{}_object_mesh_{}.vtp",
-                                                SRC_DIR, output_file_stem, output_file_stem, i);
 
     fmt::print("----- Test phase {} -----\n", i);
-
     // 1. test files exist: rank mesh, object mesh, png
     ASSERT_TRUE(
       std::filesystem::exists(rank_mesh_file)
-    ) << fmt::format("Error: rank mesh not generated at {}", object_mesh_file);
+    ) << fmt::format("Error: rank mesh not generated at {}", rank_mesh_file);
     ASSERT_TRUE(
       std::filesystem::exists(object_mesh_file)
     ) << fmt::format("Error: object mesh not generated at {}", object_mesh_file);
-    ASSERT_TRUE(
+    ASSERT_FALSE(
       std::filesystem::exists(png_file)
     ) << fmt::format("Error: PNG image not generated at {}", png_file);
-
-    // 2. test PNG with tolerance
-    fmt::print("-- Test PNG --\n");
-    auto expected_png_file = fmt::format("{}/tests/expected/{}/{}{}.png", SRC_DIR, output_file_stem, output_file_stem, i);
-    std::vector<std::string> cmd_vars = {
-      fmt::format("ACTUAL={}", png_file),
-      fmt::format("EXPECTED={}", expected_png_file),
-      "TOLERANCE=2.0",
-    };
-    auto cmd = fmt::format("{} {}/tests/test_image.sh",
-      fmt::join(cmd_vars, " "),
-      SRC_DIR
-    );
-    const auto [status, output] = Util::exec(cmd.c_str());
-    ASSERT_EQ(status, EXIT_SUCCESS) << fmt::format("Error: {}", output);
-
-    // TODO: testing mesh files cannot be a simple diff as below because each run generates some different data.
-    //       The future test should test XML Nodes
-    // 3. test vtp's file content
-    // 3.1 rank mesh file
-    fmt::print("-- Test rank mesh --\n");
-
-    // auto rank_mesh_content = Util::getFileContent(rank_mesh_file);
-    // auto expected_rank_mesh_content = Util::getFileContent(expected_rank_mesh_file);
-    // ASSERT_EQ(expected_rank_mesh_content, rank_mesh_content) << fmt::format("rank mesh file content differs from expected at phase {}", i);
-
-    vtkNew<vtkXMLPolyDataReader> expected_rank_mesh_reader;
-    expected_rank_mesh_reader->SetFileName(expected_rank_mesh_file.c_str());
-    expected_rank_mesh_reader->Update();
-    vtkPolyData *expected_rank_mesh = expected_rank_mesh_reader->GetOutput();
-
-    vtkNew<vtkXMLPolyDataReader> rank_mesh_reader;
-    rank_mesh_reader->SetFileName(rank_mesh_file.c_str());
-    rank_mesh_reader->Update();
-    vtkPolyData *rank_mesh = rank_mesh_reader->GetOutput();
-
-    this->assertPolyEquals(rank_mesh, expected_rank_mesh);
-
-    // 3.2 object mesh file
-    fmt::print("-- Test object mesh --\n");
-
-    // auto object_mesh_content = Util::getFileContent(object_mesh_file);
-    // auto expected_object_mesh_content = Util::getFileContent(expected_object_mesh_file);
-    // ASSERT_EQ(expected_object_mesh_content, object_mesh_content) << fmt::format("object mesh file content differs from expected at phase {}", i);
-
-    vtkNew<vtkXMLPolyDataReader> expected_object_mesh_reader;
-    expected_object_mesh_reader->SetFileName(expected_object_mesh_file.c_str());
-    expected_object_mesh_reader->Update();
-    vtkPolyData *expected_object_mesh = expected_object_mesh_reader->GetOutput();
-
-    vtkNew<vtkXMLPolyDataReader> object_mesh_reader;
-    object_mesh_reader->SetFileName(object_mesh_file.c_str());
-    object_mesh_reader->Update();
-    vtkPolyData *object_mesh = object_mesh_reader->GetOutput();
-
-    this->assertPolyEquals(object_mesh, expected_object_mesh);
-
-    fmt::print("----- End Test phase {} -----\n", i);
 
     } // end phases loop
 }
@@ -258,8 +118,8 @@ INSTANTIATE_TEST_SUITE_P(
     ParseRenderTests,
     ParseRenderTest,
     ::testing::Values<std::string>(
-        "conf.yaml",
-        "ccm-example.yaml"
+        "conf-no-png.yaml",
+        "ccm-example-no-png.yaml"
     ),
     [](const ::testing::TestParamInfo<std::string>& in_info) {
       // test suffix as slug
